LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/drawing-rectangles-to-the-canvas/2d.strokeRect.clip-expected.txt

2d.strokeRect.clip

strokeRect is affected by clipping regions


PASS strokeRect is affected by clipping regions
